Agregar RestauranteTest y actualizar logica de Restaurante

+++ src/com/restaurante/Restaurante.java
@@ -4,7 +4,7 @@ public class Restaurante {
 	
 	public String nombre;
 	public String direccion;
-	public float calificacion;
+	public double calificacion;
 
 
 }
